+++ composeApp/src/commonMain/kotlin/ai/mcpdirect/studio/app/home/widget/MCPServersWidget.kt
@@ -190,7 +190,7 @@ fun TeamToolMakerCard(
             IconButton(
                 onClick = {
                     generalViewModel.currentScreen(
-                        Screen.MCPTeam(team,toolMaker),
+                        Screen.MCPTeam(team,listOf(toolMaker)),
                         "My Teams",
                         Screen.Home
                     )